[TS-326] Refactor: my info page refactoring
